more code restructuring
