+++ frontend/src/lib/components/news/create-news-form/components/news-form-fields/news-header-image-form-field/NewsHeaderImageFormField.tsx
